Update README.md
This Bank System project was developed as part of completing Course #11 in Programming Advices. It is intended purely for study and educational purposes, demonstrating fundamental banking operations such as creating accounts, tracking balances, and handling transactions in a simplified environment.

The project is open source, so feel free to explore, learn from, modify, or improve the code.